fix the type for the double

+++ SnakesOnAPhone_app/src/com/example/android/snake/SnakeView.java
@@ -69,7 +69,7 @@ public class SnakeView extends TileView {
 
 	// Modify these two parameters to make the game more challenging if you like.
 	private static final int INITIAL_MOVE_DELAY = 300;
-	private static final float MOVE_DELAY_MULTIPLIER = 0.85;
+	private static final double MOVE_DELAY_MULTIPLIER = 0.85;
 	
     /**
      * mScore: used to track the number of apples captured mMoveDelay: number of
